renamed file to match util function renaming
