+++ src/utils/append-trailing-punctuation.js
